pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "simple-smartsheet"
version = "0.3.0"
description = "Python library to interact with Smartsheet API"
authors = ["Dmitry Figol <git@dmfigol.me>"]
license = "MIT"

readme = "README.md"

homepage = "https://github.com/dmfigol/simple-smartsheet"
repository = "https://github.com/dmfigol/simple-smartsheet"

keywords = ["smartsheet"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
]

[tool.poetry.dependencies]
python = ">=3.6"
marshmallow = "^3.0.0rc1"
requests = "*"
attrs = "*"
cattrs = "*"
mypy-extensions = "*"
aiohttp = "*"

[tool.poetry.dev-dependencies]
flake8 = "*"
mypy = "*"
black = { version = "*", allows-prereleases = true }
isort = "*"
ipython = "*"
ipdb = "*"
pytest = "*"
